// src/models/routes/screenName.enum.ts
enum ScreenNameEnum {
  //Home Screen
  SPLASH_SCREEN = 'SPLASH_SCREEN',
  HOME_TAB_SCREEN = 'HOME_TAB_SCREEN',
  LOGIN_SCREEN = 'LOGIN_SCREEN',

  ADD_PLANNER_SCREEN = 'ADD_PLANNER_SCREEN',
  PLANNER_DETAIL_SCREEN = 'PLANNER_DETAIL_SCREEN',
  SELECT_LANGUAGE_SCREEN = 'SELECT_LANGUAGE_SCREEN',

  CTM_SCREEN = 'CTM_SCREEN',
  //Tabs
  PLANNER_LIST_SCREEN = 'PLANNER_LIST_SCREEN',
  ADD_SERVICE_SCREEN = 'ADD_SERVICE_SCREEN',
  REPORTS_SCREEN = 'REPORTS_SCREEN',
  LOGISTIC_SCREEN = 'LOGISTIC_SCREEN',
  USER_PROFILE_SCREEN = 'USER_PROFILE_SCREEN',
}

export default ScreenNameEnum;